compile issue continue...

+++ apps/include/console.h
@@ -1,7 +1,9 @@
 #ifndef __CONSOLE_H__
 #define __CONSOLE_H__
 #include<stdarg.h>
-void console_init(int uart_id, u32 baudrate);
+#include<stdio.h>
+#include<string.h>
+void console_init(int uart_id, unsigned long baudrate);
 unsigned long console_baud(void);
 void console_flush(void);
 
@@ -13,7 +15,7 @@ void console_puts(const char* s);
 int console_gets(char* buf, int len);
 
 int console_printf(const char* fmt, ...);
-
+// int printf(const char* fmt, ...);
 // int console_check_char(unsigned char check);
 
 #endif
